update: updater process
