+++ GUI/uic/qfluent/components/updater.py
@@ -1,12 +1,17 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 import re
-from PyQt5.QtCore import Qt
+
+from PyQt5.QtCore import Qt, QTimer
+from PyQt5.QtWidgets import QApplication
 from qfluentwidgets import (
-    MessageBoxBase, TextBrowser, SubtitleLabel, StateToolTip
+    MessageBoxBase, TextBrowser, SubtitleLabel, StateToolTip, PrimaryPushButton, FluentIcon as FIF
 )
+
 from assets import res
+from utils import code_env
 from utils.docs import MarkdownConverter
+from GUI.uic.qfluent.components.cust import CustomInfoBar
 
 
 class UpdaterMessageBox(MessageBoxBase):
@@ -24,9 +29,18 @@ class UpdaterMessageBox(MessageBoxBase):
         self.widget.setMinimumWidth(int(parent.width() * 0.8))
 
     def validate(self):
-        self.gui.updaterStateTooltip = StateToolTip("Updating", res.Updater.doing, self.gui.textBrowser)
-        self.gui.updaterStateTooltip.show()
-        self.gui.conf_dia.puThread.update_signal.emit()
+        if code_env == "uv":
+            copy_command_btn = PrimaryPushButton(FIF.COPY, res.Updater.uv_update_command_btn, self)
+            copy_command_btn.clicked.connect(self.copy_command)
+            CustomInfoBar.show_custom("", res.Updater.uv_update_desc, self.gui.textBrowser, 
+                widgets=[copy_command_btn], _type="INFORMATION")
+        elif code_env == "git":
+            CustomInfoBar.show_custom("", res.Updater.git_update_desc, self.gui.textBrowser, _type="INFORMATION")
+            QTimer.singleShot(3000, self.gui.conf_dia.puThread.update_signal.emit)
+        else:
+            self.gui.updaterStateTooltip = StateToolTip("Updating", res.Updater.doing, self.gui.textBrowser)
+            self.gui.updaterStateTooltip.show()
+            self.gui.conf_dia.puThread.update_signal.emit()
         return True
 
     def show_release_note(self, note):
@@ -37,3 +51,11 @@ class UpdaterMessageBox(MessageBoxBase):
         self.textBrowser.setHtml(html_text)
         self.gui.conf_dia.hide()
         self.show()
+
+    def copy_command(self):
+        clipboard = QApplication.clipboard()
+        cmd = "uv tool upgrade ComicGUISpider"
+        if res.lang == "zh_CN":
+            cmd += " --index-url https://pypi.tuna.tsinghua.edu.cn/simple"
+        clipboard.setText(cmd)
+        self.gui.conf_dia.puThread.update_signal.emit()
